pyproject.toml
[tool.poetry]
name = "ojs-to-oaktrust"
version = "0.1.0"
description = ""
authors = ["Mark Baggett <markpbaggett@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyyaml = "^6.0.2"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
